Board Entity, dto 작성 완료
