+++ src/main/kotlin/com/b2/backoffice/domain/board/dto/BoardCreateRequest.kt
@@ -1,5 +1,6 @@
 package com.b2.backoffice.domain.board.dto
 
 data class BoardCreateRequest(
-    val a:String
+    var title : String,
+    var contents : String,
 )
