+++ src/main/kotlin/com/b2/backoffice/domain/board/dto/BoardUpdateRequest.kt
@@ -1,5 +1,7 @@
 package com.b2.backoffice.domain.board.dto
 
 data class BoardUpdateRequest(
-    val a : String
+    val password : String,
+    var title : String,
+    var contents : String,
 )
